+++ MonoEngine/Rendering/IRenderBuffer.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "System/SysTypes.h"
+#include <cstddef>
 
 namespace mono
 {
@@ -12,9 +12,9 @@ namespace mono
         virtual ~IRenderBuffer()
         { }
 
-        virtual void UpdateData(const void* data, uint offset, uint count) = 0;
+        virtual void UpdateData(const void* data, size_t offset, size_t count) = 0;
         virtual void Use() const = 0;
-        virtual uint Id() const = 0;
+        virtual size_t Id() const = 0;
     };
 
     void ClearRenderBuffer();
