+++ MonoEngine/Rendering/RenderBuffer.cpp
@@ -4,11 +4,11 @@
 
 using namespace mono;
 
-RenderBuffer::RenderBuffer(BufferType type, BufferData data, uint count)
+RenderBuffer::RenderBuffer(BufferType type, BufferData data, size_t count)
     : m_type(type),
       m_data(data)
 {
-    const uint size = (m_data == BufferData::FLOAT) ? sizeof(float) : sizeof(int);
+    const size_t size = (m_data == BufferData::FLOAT) ? sizeof(float) : sizeof(int);
     const GLenum usage = (m_type == BufferType::STATIC) ? GL_STATIC_DRAW : GL_DYNAMIC_DRAW;
 
     glGenBuffers(1, &m_buffer_id);
@@ -21,9 +21,9 @@ RenderBuffer::~RenderBuffer()
     glDeleteBuffers(1, &m_buffer_id);
 }
 
-void RenderBuffer::UpdateData(const void* data, uint offset, uint count)
+void RenderBuffer::UpdateData(const void* data, size_t offset, size_t count)
 {
-    const uint size = (m_data == BufferData::FLOAT) ? sizeof(float) : sizeof(int);
+    const size_t size = (m_data == BufferData::FLOAT) ? sizeof(float) : sizeof(int);
 
     Use();
     glBufferSubData(GL_ARRAY_BUFFER, offset * size, size * count, data);
@@ -34,7 +34,7 @@ void RenderBuffer::Use() const
     glBindBuffer(GL_ARRAY_BUFFER, m_buffer_id);
 }
 
-uint RenderBuffer::Id() const
+size_t RenderBuffer::Id() const
 {
     return m_buffer_id;
 }
